Se agrega middlewares para validar regstro de ususarios y multer para registro de usuarios
