+++ public/css/users.css
@@ -97,8 +97,24 @@ main h2{
     margin: 15px;
 }
 
+is-invalid:focus{
+    border-color:#dc3545;
+    box-shadow:0 0 0 .25rem rgba(220,53,69,.25);
+    padding-right:calc(1.5em + .75rem);
+    background-position:top calc(.375em + .1875rem) right calc(.375em + .1875rem);
+    
+    background-position:right .75rem center,center right 1.75rem;
+    background-size:16px 12px,calc(.75em + .375rem) calc(.75em + .375rem);
+
+
+    border-color:#dc3545;
+    box-shadow:0 0 0 .25rem rgba(220,53,69,.25)
+}
+
+
+
 .text-danger{
-    color: rgb(248, 66, 66);
+    color: rgb(248, 66, 66);    
     margin-top: 5px;
 }
 /* Fin de Formulario de edicion de login */
