+++ views/partials/head.ejs
@@ -13,4 +13,5 @@
     <link rel="stylesheet" href="/css/home.css">
     <link rel="stylesheet" href="/css/products.css">
     <link rel="stylesheet" href="/css/users.css">
+    <!-- <link rel="stylesheet" href="/css/bootstrap.min.css"> -->
 </head>
